requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.27
alembic>=1.13.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
openai>=1.12.0
pydantic[email]>=2.6.0
python-multipart>=0.0.9
aiosqlite>=0.20.0
pydantic-settings>=2.0.0
psycopg2-binary>=2.9.0
